refactored goals.py into smaller files
